third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/workers/cross-origin-unsupported-url-expected.txt

CONSOLE ERROR: Failed to load ftp://127.0.0.1/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load ftp://127.0.0.1/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: Failed to load ftp://127.0.0.1/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load ftp://127.0.0.1/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: Failed to load localhost:8080/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load localhost:8080/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: Failed to load localhost:8080/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load localhost:8080/: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: Failed to load tel:1234: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load tel:1234: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: Failed to load tel:1234: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
CONSOLE ERROR: line 1: Failed to load tel:1234: Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
[Worker] Test cross-origin XHRs to CORS-unsupported protocol schemes in the URL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: ../resources/cross-origin-unsupported-url.js
PASS [Worker] errorEvent.name is "NetworkError"
PASS [Worker] errorEvent.type is "error"
PASS [Worker] xhr.send() did not throw exception.
PASS [Worker] errorEvent.name is "NetworkError"
PASS [Worker] errorEvent.type is "error"
PASS [Worker] xhr.send() did not throw exception.
PASS [Worker] errorEvent.name is "NetworkError"
PASS [Worker] errorEvent.type is "error"
PASS [Worker] xhr.send() did not throw exception.
PASS [Worker] errorEvent.name is "NetworkError"
PASS [Worker] errorEvent.type is "error"
PASS [Worker] xhr.send() did not throw exception.
PASS [Worker] errorEvent.name is "NetworkError"
PASS [Worker] errorEvent.type is "error"
PASS [Worker] xhr.send() did not throw exception.
PASS [Worker] errorEvent.name is "NetworkError"
PASS [Worker] errorEvent.type is "error"
PASS [Worker] xhr.send() did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
